ddd/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "app"
authors = ["Sergey Bilibin"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
redis="==4.4.2"
aiohttp="==3.8.5"
elasticsearch="==7.17.0"
fastapi=">=0.68.0"
orjson="==3.4.7"
pydantic="==1.9.0"
uvicorn="==0.12.2"
uvloop="==0.17.0"
python-dotenv="==1.0.0"
gunicorn="==20.1.0"
httptools="==0.5.0"
httpx="==0.23.3"
dependency-injector="==4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"